+++ ContentPatcher/Framework/Tokens/ValueProviders/Players/LocalOrHostPlayerValueProvider.cs
@@ -4,7 +4,6 @@ using System.Linq;
 using ContentPatcher.Framework.Conditions;
 using ContentPatcher.Framework.Constants;
 using Pathoschild.Stardew.Common.Utilities;
-using StardewModdingAPI;
 using StardewValley;
 
 namespace ContentPatcher.Framework.Tokens.ValueProviders.Players
@@ -25,11 +24,9 @@ namespace ContentPatcher.Framework.Tokens.ValueProviders.Players
         private readonly Func<Farmer, InvariantHashSet> FetchValues;
 
         /// <summary>The values as of the last context update.</summary>
-        private readonly IDictionary<PlayerType, InvariantHashSet> Values = new Dictionary<PlayerType, InvariantHashSet>
-        {
-            [PlayerType.CurrentPlayer] = new(),
-            [PlayerType.HostPlayer] = new()
-        };
+        private readonly IDictionary<long, InvariantHashSet> Values = new Dictionary<long, InvariantHashSet>();
+
+        private readonly IDictionary<PlayerType, long> PlayerIdsByType = new Dictionary<PlayerType, long>();
 
 
         /*********
@@ -66,23 +63,45 @@ namespace ContentPatcher.Framework.Tokens.ValueProviders.Players
             {
                 bool changed = false;
 
-                foreach (var pair in this.Values)
+                if (this.MarkReady(this.SaveReader.IsReady))
                 {
-                    PlayerType type = pair.Key;
-                    InvariantHashSet values = pair.Value;
-                    InvariantHashSet oldValues = new InvariantHashSet(values);
-
-                    values.Clear();
-
-                    if (this.MarkReady(this.SaveReader.IsReady))
+                    // update player type => ID map
+                    foreach (PlayerType type in Enum.GetValues(typeof(PlayerType)))
                     {
-                        Farmer player = this.SaveReader.GetPlayer(type);
-
-                        if (player != null)
-                            values.AddMany(this.FetchValues(player));
+                        long id = this.SaveReader.GetPlayer(type).UniqueMultiplayerID;
+                        changed |= !this.PlayerIdsByType.TryGetValue(type, out long oldId) || oldId != id;
+                        this.PlayerIdsByType[type] = id;
                     }
 
-                    changed |= this.IsChanged(oldValues, values);
+                    // update values by player ID
+                    HashSet<long> removeIds = new HashSet<long>(this.Values.Keys);
+                    foreach (Farmer player in this.SaveReader.GetAllPlayers())
+                    {
+                        // get values
+                        long id = player.UniqueMultiplayerID;
+                        InvariantHashSet newValues = this.FetchValues(player);
+                        if (!this.Values.TryGetValue(id, out InvariantHashSet oldValues))
+                            oldValues = null;
+
+                        // track changes
+                        removeIds.Remove(id);
+                        changed |= oldValues == null || this.IsChanged(oldValues, newValues);
+
+                        // update
+                        this.Values[id] = newValues;
+                    }
+
+                    // remove players if needed
+                    foreach (long id in removeIds)
+                    {
+                        this.Values.Remove(id);
+                        changed = true;
+                    }
+                }
+                else
+                {
+                    this.Values.Clear();
+                    this.PlayerIdsByType.Clear();
                 }
 
                 return changed;
@@ -107,8 +126,8 @@ namespace ContentPatcher.Framework.Tokens.ValueProviders.Players
         {
             this.AssertInput(input);
 
-            return this.TryParseInput(input, out ISet<PlayerType> playerTypes, out _)
-                ? this.GetValuesFor(playerTypes)
+            return this.TryParseInput(input, out ISet<long> playerIds, out _)
+                ? this.GetValuesFor(playerIds)
                 : Enumerable.Empty<string>();
         }
 
@@ -118,41 +137,57 @@ namespace ContentPatcher.Framework.Tokens.ValueProviders.Players
         *********/
         /// <summary>Parse the input arguments if valid.</summary>
         /// <param name="input">The input arguments.</param>
-        /// <param name="playerTypes">The parsed player types.</param>
+        /// <param name="playerIds">The parsed player IDs.</param>
         /// <param name="error">The error indicating why the input is invalid, if applicable.</param>
         /// <returns>Returns whether the input is valid.</returns>
-        private bool TryParseInput(IInputArguments input, out ISet<PlayerType> playerTypes, out string error)
+        private bool TryParseInput(IInputArguments input, out ISet<long> playerIds, out string error)
         {
-            playerTypes = new HashSet<PlayerType>();
+            playerIds = new HashSet<long>();
+
             foreach (string arg in input.PositionalArgs)
             {
-                if (!Enum.TryParse(arg, ignoreCase: true, out PlayerType type))
+                if (Enum.TryParse(arg, ignoreCase: true, out PlayerType type))
                 {
-                    error = $"invalid input arguments ({input.TokenString}) for {this.Name} token, expected any of {string.Join(", ", this.Values.Keys.Select(p => p.ToString()))}.";
+                    if (this.PlayerIdsByType.TryGetValue(type, out long id))
+                        playerIds.Add(id);
+                }
+                else if (long.TryParse(arg, out long id))
+                    playerIds.Add(id);
+                else
+                {
+                    error = $"invalid input arguments ({input.TokenString}) for {this.Name} token, expected any combination of player IDs, {string.Join(", ", Enum.GetNames(typeof(PlayerType)))}.";
                     return false;
                 }
-
-                playerTypes.Add(type);
             }
 
             error = null;
             return true;
         }
 
-        /// <summary>Get the cached values for the given player types.</summary>
-        /// <param name="playerTypes">The player types.</param>
-        private IEnumerable<string> GetValuesFor(ISet<PlayerType> playerTypes)
+        /// <summary>Get the cached values for the given player IDs.</summary>
+        /// <param name="playerIds">The player IDs.</param>
+        private IEnumerable<string> GetValuesFor(ISet<long> playerIds)
         {
-            // current player
-            if (playerTypes.Count == 0 || Context.IsMainPlayer)
-                return this.Values[PlayerType.CurrentPlayer];
+            // default to current player
+            if (!playerIds.Any())
+                playerIds.Add(Game1.player.UniqueMultiplayerID);
 
-            // one player type
-            if (playerTypes.Count == 1)
-                return this.Values[playerTypes.First()];
+            // get single value (avoids copying the collection in most cases)
+            if (playerIds.Count == 1)
+            {
+                return this.Values.TryGetValue(playerIds.First(), out InvariantHashSet set)
+                    ? set
+                    : Enumerable.Empty<string>();
+            }
 
-            // multiple player types
-            return new InvariantHashSet(playerTypes.SelectMany(type => this.Values[type]));
+            // get multiple values
+            HashSet<string> values = new();
+            foreach (long id in playerIds)
+            {
+                if (this.Values.TryGetValue(id, out InvariantHashSet set))
+                    values.AddMany(set);
+            }
+            return values;
         }
     }
 }